+++ drake/examples/kuka_iiwa_arm/controlled_kuka/kuka_demo.cc
@@ -22,7 +22,7 @@
 #include "drake/systems/trajectories/piecewise_polynomial_trajectory.h"
 
 // Includes for the planner.
-#include "drake/multibody/IKoptions.h"
+#include "drake/multibody/ik_options.h"
 #include "drake/multibody/RigidBodyIK.h"
 
 DEFINE_double(simulation_sec, 0.5, "Number of seconds to simulate.");
