+++ drake/matlab/systems/plants/IKoptionsmex.cpp
@@ -1,7 +1,7 @@
 #include <mex.h>
 
 #include "drake/matlab/util/drakeMexUtil.h"
-#include "drake/multibody/IKoptions.h"
+#include "drake/multibody/ik_options.h"
 #include "drake/multibody/rigid_body_tree.h"
 
 using namespace std;
