+++ drake/matlab/systems/plants/approximateIKmex.cpp
@@ -3,7 +3,7 @@
 #include "drake/multibody/RigidBodyIK.h"
 #include "drake/multibody/rigid_body_tree.h"
 #include "drake/multibody/constraint/RigidBodyConstraint.h"
-#include "drake/multibody/IKoptions.h"
+#include "drake/multibody/ik_options.h"
 #include "drake/matlab/util/drakeMexUtil.h"
 
 using namespace std;
