+++ drake/matlab/systems/plants/inverseKinTrajmex.cpp
@@ -2,7 +2,7 @@
 
 #include "drake/multibody/rigid_body_tree.h"
 #include "drake/multibody/constraint/RigidBodyConstraint.h"
-#include "drake/multibody/IKoptions.h"
+#include "drake/multibody/ik_options.h"
 #include "drake/multibody/RigidBodyIK.h"
 #include <Eigen/Dense>
 #include "drake/matlab/util/drakeMexUtil.h"
