+++ drake/multibody/approximateIK.cpp
@@ -1,7 +1,7 @@
 #include "drake/multibody/RigidBodyIK.h"
 #include "drake/multibody/rigid_body_tree.h"
 #include "drake/multibody/constraint/RigidBodyConstraint.h"
-#include "drake/multibody/IKoptions.h"
+#include "drake/multibody/ik_options.h"
 #include <math.h>
 #include <set>
 #include <gurobi_c++.h>
