+++ drake/multibody/test/testApproximateIK.cpp
@@ -1,6 +1,6 @@
 #include "drake/multibody/RigidBodyIK.h"
 #include "drake/multibody/rigid_body_tree.h"
-#include "drake/multibody/IKoptions.h"
+#include "drake/multibody/ik_options.h"
 #include "drake/multibody/constraint/RigidBodyConstraint.h"
 #include <iostream>
 #include <cstdlib>
